requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
paho-mqtt>=1.6.1
toml>=0.10.2
